cohort-1/track-2/agroadvisor/requirements.txt
google-adk>=0.3.0
flask>=3.0
httpx>=0.27
mcp>=1.0
pytest>=8.0
pytest-asyncio>=0.23
gunicorn>=22.0
